It's all just so fucked
1. accept should be blocking; need to think of a way to implement that or develop a means of rationalizing making it not blocking.
- the issue is that a) if the syn comes in before the acceopt, this approach is okay
- if the syn comes in after the accept, we are in trouble because we need to set a state variable

However, the bigger problem is the semantics.

accept is supposed to
1. be blocking
2. return the newly created socket for listening on the 4-tuple.

our accept is a) nonblocking (will return immediately regardless of whether the underlying thread has actually accepted) and b) returns the original socket. This is because we create new TCPSocks in threads (because we did not think far enough to use selector/multiplexer) so we are stuck with TCP connections in individual threads. To implement this feature requires the implementation of a blocking accept; without a blocking accept, you will not necessarily be able to return the copleted socket. \

The "slack" -- delay for receiving a message is picked up by the server code, where we use an infinite loop and the write function that can safely return 0 if no bytes can be read.
We will effectively return 0 until the connection is fully accepted on the listensock//now mp connection 1